Added ability to run a Ruby file before any children are forked (useful for pre-loading libraries) using --prerun [FILE]
Altered the process name to include the command line arguments too. This is
because I can't work out how to set the process cmdline to a string that is
shorter than the 'real' cmdline in Ruby and it was showing blank newlines when
Jobby was called with Long argument lists.

Built the 0.2.0 gem.